Order service implementation is added to home assignment.
Save order and retreive order operations are provided by system.
Order and Product Many-to-Many relation is only defined in Order entity.

TODO:

OrderRequestDTO can be simplified and can keep OrderItems (SKU and amount of product).
Better LocalDateTime conversion is needed for retreive operation of OrderController.

Some ToDos are to code :)

Default time zone is set to EST. This can be moved to application properties and can be configured.

Better integration tests needed :-m